Segregate host entities
